# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CareTrack API Documentation",
    "description": "API documentation for the CareTrack backend application. To access protected endpoints, click the 'Authorize' button and enter your JWT token.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
  "paths": {
    "/medical-records/{recordId}": {
      "get": {
        "tags": [
          "medical-record-controller"
        ],
        "operationId": "getRecordById",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecord"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "medical-record-controller"
        ],
        "operationId": "updateMedicalRecord",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecord"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecord"
                }
              }
            }
          }
        }
      }
    },
...
    "/medical-records/share": {
      "post": {
        "tags": [
          "medical-record-controller"
        ],
        "operationId": "shareRecordWithDoctor",
        "parameters": [
          {
            "name": "recordId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/patient": {
      "get": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "getPatient",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserMainView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "patient-controller"
        ],
        "operationId": "updatePatient",
        "requestBody": {
          "description": "Dynamic map containing patient fields to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Map containing any combination of patient fields to update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserMainView"
                }
              }
            }
          }
        }
      }
    },
    "/doctor": {
      "get": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "getDoctor",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorMainView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "updateDoctor",
        "requestBody": {
          "description": "Dynamic map containing doctor fields to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Map containing any combination of doctor fields to update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorMainView"
                }
              }
            }
          }
        }
      }
    },
    "/medical-records/shared/patient/me": {
      "get": {
        "tags": [
          "medical-record-controller"
        ],
        "operationId": "getSharedRecordsForPatient",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicalRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/medical-records/shared/doctor/me": {
      "get": {
        "tags": [
          "medical-record-controller"
        ],
        "operationId": "getSharedRecordsForDoctor",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicalRecordResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/medical-records/patient/me": {
      "get": {
        "tags": [
          "medical-record-controller"
        ],
        "operationId": "getRecordsByPatient",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicalRecordResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Doctor": {
        "required": [
          "availableDays",
          "city",
          "consultationFee",
          "doctorSpeciality",
          "email",
          "endTime",
          "fullName",
          "info",
          "password",
          "patientNumber",
          "phoneNumber",
          "startTime",
          "street"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "maxLength": 50,
            "minLength": 7,
            "type": "string"
          },
          "email": {
            "pattern": "^[^\\s@]+@(gmail|yahoo|hotmail|outlook)\\.(com|net|org)$",
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 6,
            "type": "string",
            "writeOnly": true
          },
          "phoneNumber": {
            "maxLength": 11,
            "minLength": 11,
            "pattern": "^(010|011|012|015)\\d{8}$",
            "type": "string"
          },
          "city": {
            "type": "string",
            "enum": [
              "CAIRO",
              "GIZA",
              "ALEXANDRIA",
              "LUXOR",
              "ASWAN",
              "PORT_SAID",
              "SUEZ",
              "ISMAILIA",
              "FAIYUM",
              "DAMIETTA",
              "ASYUT",
              "MINYA",
              "BENI_SUEF",
              "SHARQIA",
              "DAKAHLIYA",
              "GHARBIA",
              "MONUFIA",
              "KAFR_EL_SHEIKH",
              "BEHEIRA",
              "QENA",
              "SOHAG",
              "RED_SEA",
              "MATROUH",
              "NEW_VALLEY",
              "NORTH_SINAI",
              "SOUTH_SINAI"
            ]
          },
          "street": {
            "maxLength": 2147483647,
            "minLength": 5,
            "type": "string"
          },
          "doctorSpeciality": {
            "type": "string",
            "enum": [
              "CARDIOLOGY",
              "DERMATOLOGY",
              "NEUROLOGY",
              "ORTHOPEDICS",
              "OPHTHALMOLOGY",
              "OTOLARYNGOLOGY"
            ]
          },
          "info": {
            "maxLength": 500,
            "minLength": 25,
            "type": "string"
          },
          "patientNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "endTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "consultationFee": {
            "type": "number",
            "format": "double"
          },
          "availableDays": {
            "maxItems": 7,
            "minItems": 0,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "MONDAY",
                "TUESDAY",
                "WEDNESDAY",
                "THURSDAY",
                "FRIDAY",
                "SATURDAY",
                "SUNDAY"
              ]
            }
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "accessibleMedicalRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalRecord"
            }
          },
          "medicalRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalRecord"
            }
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "PATIENT",
              "DOCTOR"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonLocked": {
            "type": "boolean"
          }
        }
      },
...
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MedicalRecord": {
        "required": [
          "content",
          "date"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "maxLength": 500,
            "minLength": 25,
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "sharedWithDoctors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Doctor"
            }
          },
          "doctor": {
            "$ref": "#/components/schemas/Doctor"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          }
        }
      },
...
      "UserMainView": {
        "required": [
          "email",
          "fullName",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "email": {
            "pattern": "^[^\\s@]+@(gmail|yahoo|hotmail|outlook)\\.(com|net|org)$",
            "type": "string"
          },
          "fullName": {
            "maxLength": 50,
            "minLength": 7,
            "type": "string"
          },
          "phoneNumber": {
            "maxLength": 11,
            "minLength": 11,
            "pattern": "^(010|011|012|015)\\d{8}$",
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "PATIENT",
              "DOCTOR"
            ]
          }
        }
      },
...
      "DoctorMainView": {
        "required": [
          "availableDays",
          "city",
          "consultationFee",
          "doctorSpeciality",
          "email",
          "endTime",
          "fullName",
          "info",
          "patientNumber",
          "phoneNumber",
          "startTime",
          "street"
        ],
        "type": "object",
        "properties": {
          "email": {
            "pattern": "^[^\\s@]+@(gmail|yahoo|hotmail|outlook)\\.(com|net|org)$",
            "type": "string"
          },
          "fullName": {
            "maxLength": 50,
            "minLength": 7,
            "type": "string"
          },
          "phoneNumber": {
            "maxLength": 11,
            "minLength": 11,
            "pattern": "^(010|011|012|015)\\d{8}$",
            "type": "string"
          },
          "city": {
            "type": "string",
            "enum": [
              "CAIRO",
              "GIZA",
              "ALEXANDRIA",
              "LUXOR",
              "ASWAN",
              "PORT_SAID",
              "SUEZ",
              "ISMAILIA",
              "FAIYUM",
              "DAMIETTA",
              "ASYUT",
              "MINYA",
              "BENI_SUEF",
              "SHARQIA",
              "DAKAHLIYA",
              "GHARBIA",
              "MONUFIA",
              "KAFR_EL_SHEIKH",
              "BEHEIRA",
              "QENA",
              "SOHAG",
              "RED_SEA",
              "MATROUH",
              "NEW_VALLEY",
              "NORTH_SINAI",
              "SOUTH_SINAI"
            ]
          },
          "street": {
            "maxLength": 2147483647,
            "minLength": 5,
            "type": "string"
          },
          "doctorSpeciality": {
            "type": "string",
            "enum": [
              "CARDIOLOGY",
              "DERMATOLOGY",
              "NEUROLOGY",
              "ORTHOPEDICS",
              "OPHTHALMOLOGY",
              "OTOLARYNGOLOGY"
            ]
          },
          "info": {
            "maxLength": 500,
            "minLength": 25,
            "type": "string"
          },
          "patientNumber": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "endTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "consultationFee": {
            "type": "number",
            "format": "double"
          },
          "availableDays": {
            "maxItems": 7,
            "minItems": 0,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "MONDAY",
                "TUESDAY",
                "WEDNESDAY",
                "THURSDAY",
                "FRIDAY",
                "SATURDAY",
                "SUNDAY"
              ]
            }
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "PATIENT",
              "DOCTOR"
            ]
          }
        }
      },
...
      "MedicalRecordResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorName": {
            "type": "string"
          },
          "patientId": {
            "type": "integer",
            "format": "int64"
          },
          "patientName": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "description": "Enter your JWT token in the format: your-jwt-token",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}